+++ jest.base.js
@@ -1,0 +1,23 @@
+module.exports = {
+    transform: {
+        '^.+\\.ts$': 'ts-jest'
+    },
+    collectCoverage: true,
+    collectCoverageFrom: ['src/**/*.ts'],
+    coverageReporters: [
+        'text', 
+        ['lcov', { projectRoot: '../../' }]
+    ],
+    reporters: [
+        'default',
+        [
+            'jest-sonar',
+            {
+                reportedFilePath: 'relative',
+                relativeRootDir: '<rootDir>/../../../'
+            }
+        ]
+    ],
+    modulePathIgnorePatterns: ['<rootDir>/dist', '<rootDir>/coverage'],
+    verbose: true
+};