+++ packages/axios-extension/src/abap/adt-catalog/services/index.ts
@@ -1,4 +1,4 @@
 export { AdtService } from './adt-service';
 export { AtoService } from './ato-service';
 export { TransportChecksService } from './transportcheck-service';
-export { TransportRequestService } from './transportrequest-service';
+export { TransportRequestService, NewUi5ObjectRequestParams } from './transportrequest-service';