+++ packages/axios-extension/src/abap/ui5-abap-repository-service.ts
@@ -1,7 +1,62 @@
-import type { AxiosResponse, AxiosRequestConfig, AxiosError } from 'axios';
-import { readFileSync } from 'fs';
+import type { AxiosResponse, AxiosRequestConfig } from 'axios';
 import { prettyPrintError, prettyPrintMessage } from './message';
 import { ODataService } from '../base/odata-service';
+import { isAxiosError } from '../base/odata-request-error';
+
+/**
+ * Required configuration for the BSP hosting an app.
+ */
+export interface BspConfig {
+    /**
+     * Name of the BSP, additionally, the last part of the exposed service path
+     */
+    name: string;
+
+    /**
+     * Optional description of the ABAP development object representing the BSP
+     */
+    description?: string;
+
+    /**
+     * Optional package for the ABAP development object
+     */
+    package?: string;
+
+    /**
+     * Optional transport request to record the changes
+     */
+    transport?: string;
+}
+
+/**
+ * Configuration required for deploying an app.
+ */
+export interface DeployConfig {
+    /**
+     * archive zip archive containing the application files as buffer
+     */
+    archive: Buffer;
+    /**
+     * app application configuration
+     */
+    bsp: BspConfig;
+    /**
+     * if set to true, all requests will be sent, the service checks them, but no actual deployment will happen
+     */
+    testMode?: boolean;
+    /**
+     * if set then the SafeMode url parameter will be set. SafeMode is by default active, to deactivate provide false
+     */
+    safeMode?: boolean;
+}
+
+/**
+ * Configuration required for undeploying an app.
+ */
+export interface UndeployConfig {
+    bsp: Pick<BspConfig, 'name' | 'transport'>;
+    testMode?: DeployConfig['testMode'];
+}
 
 /**
  * Application information object returned by the UI5 Repository service
@@ -12,16 +67,6 @@ export interface AppInfo {
     Description?: string;
     Info?: string;
     ZipArchive?: string;
-}
-
-/**
- * Required configuration to deploy an application using the UI5 Repository service.
- */
-export interface ApplicationConfig {
-    name: string;
-    description?: string;
-    package?: string;
-    transport?: string;
 }
 
 export const abapUrlReplaceMap = new Map([
@@ -56,6 +101,19 @@ export class Ui5AbapRepositoryService extends ODataService {
     public static readonly PATH = '/sap/opu/odata/UI5/ABAP_REPOSITORY_SRV';
 
     /**
+     * Extension of the base constructor to set preferred response format if not provided by caller.
+     *
+     * @param config optional base configuration for Axios
+     */
+    public constructor(config?: AxiosRequestConfig) {
+        config = config ?? {};
+        config.headers = config.headers ?? {};
+        // @see https://axios-http.com/docs/config_defaults
+        config.headers['Accept'] = config.headers['Accept'] ?? 'application/json,application/xml,text/plain,*/*';
+        super(config);
+    }
+
+    /**
      * Get information about a deployed application. Returns undefined if the application cannot be found.
      *
      * @param app application id (BSP application name)
@@ -73,30 +131,32 @@ export class Ui5AbapRepositoryService extends ODataService {
     /**
      * Deploy the given archive either by creating a new BSP or updating an existing one.
      *
-     * @param archivePath path to a zip archive containing the application files
-     * @param app application configuration
-     * @param testMode if set to true, all requests will be sent, the service checks them, but no actual deployment will happen
+     * @param config deployment config
+     * @param config.archive zip archive containing the application files as buffer
+     * @param config.bsp BSP configuration
+     * @param config.testMode if set to true, all requests will be sent, the service checks them, but no actual deployment will happen
+     * @param config.safeMode if set then the SafeMode url parameter will be set. SafeMode is by default active, to activate provide false
      * @returns the Axios response object for futher processing
      */
-    public async deploy(archivePath: string, app: ApplicationConfig, testMode = false): Promise<AxiosResponse> {
-        const info: AppInfo = await this.getInfo(app.name);
+    public async deploy({ archive, bsp, testMode = false, safeMode }: DeployConfig): Promise<AxiosResponse> {
+        const info: AppInfo = await this.getInfo(bsp.name);
         const payload = this.createPayload(
-            archivePath,
-            app.name,
-            app.description || 'Deployed with SAP Fiori tools',
-            info ? info.Package : app.package
+            archive,
+            bsp.name,
+            bsp.description || 'Deployed with SAP Fiori tools',
+            info ? info.Package : bsp.package
         );
-        const config = this.createConfig(app.transport, testMode);
-        const frontendUrl = this.getAbapFrontendUrl(this.defaults.baseURL);
+        this.log.debug(`Payload:\n${payload}`);
+        const config = this.createConfig(bsp.transport, testMode, safeMode);
+        const frontendUrl = this.getAbapFrontendUrl();
         try {
-            const response: AxiosResponse | undefined = await this.updateRepoRequest(!!info, app.name, payload, config);
+            const response: AxiosResponse | undefined = await this.updateRepoRequest(!!info, bsp.name, payload, config);
             // An app can be successfully deployed after a timeout exception, no value in showing exception headers
             if (response?.headers?.['sap-message']) {
-                const message = JSON.parse(response.headers['sap-message']);
-                prettyPrintMessage({ msg: message, log: this.log, host: frontendUrl });
+                prettyPrintMessage({ msg: response.headers['sap-message'], log: this.log, host: frontendUrl });
             }
             // log url of created/updated app
-            const path = '/sap/bc/ui5_ui5' + (!app.name.startsWith('/') ? '/sap/' : '') + app.name.toLowerCase();
+            const path = '/sap/bc/ui5_ui5' + (!bsp.name.startsWith('/') ? '/sap/' : '') + bsp.name.toLowerCase();
             const query = this.defaults.params?.['sap-client']
                 ? '?sap-client=' + this.defaults.params['sap-client']
                 : '';
@@ -112,20 +172,26 @@ export class Ui5AbapRepositoryService extends ODataService {
     /**
      * Undeploy an existing app.
      *
-     * @param app application configuration
-     * @param testMode if set to true, all requests will be sent, the service checks them, but no actual deployment will happen
-     * @returns the Axios response object for futher processing
+     * @param config undeployment config
+     * @param config.bsp BSP configuration
+     * @param config.testMode if set to true, all requests will be sent, the service checks them, but no actual deployment will happen
+     * @returns the Axios response object for futher processing or undefined if no request is sent
      */
-    public async undeploy(app: ApplicationConfig, testMode = false): Promise<AxiosResponse> {
-        const config = this.createConfig(app.transport, testMode);
-        const host = this.getAbapFrontendUrl(this.defaults.baseURL);
+    public async undeploy({ bsp, testMode = false }: UndeployConfig): Promise<AxiosResponse | undefined> {
+        const config = this.createConfig(bsp.transport, testMode);
+        const host = this.getAbapFrontendUrl();
         try {
-            const response = await this.deleteRepoRequest(app.name, config);
-            if (response?.headers?.['sap-message']) {
-                const message = JSON.parse(response.headers['sap-message']);
-                prettyPrintMessage({ msg: message, log: this.log, host });
+            const info: AppInfo = await this.getInfo(bsp.name);
+            if (info) {
+                const response = await this.deleteRepoRequest(bsp.name, config);
+                if (response?.headers?.['sap-message']) {
+                    prettyPrintMessage({ msg: response.headers['sap-message'], log: this.log, host });
+                }
+                return response;
+            } else {
+                this.log.warn(`Application ${bsp.name} not found, nothing to undeploy.`);
+                return undefined;
             }
-            return response;
         } catch (error) {
             this.logError({ error, host });
             throw error;
@@ -135,31 +201,32 @@ export class Ui5AbapRepositoryService extends ODataService {
     /**
      * Translate the technical ABAP on BTP URL to the frontend URL.
      *
-     * @param technicalUrl Technical URL of the ABAP system from service keys
      * @returns url to be used in the browser.
      */
-    protected getAbapFrontendUrl(technicalUrl: string): string {
+    protected getAbapFrontendUrl(): string {
+        const url = new URL(this.defaults.baseURL);
         abapUrlReplaceMap.forEach((value, key) => {
-            technicalUrl = technicalUrl.replace(key, value);
+            url.hostname = url.hostname.replace(key, value);
         });
-        return technicalUrl;
+        return `${url.protocol}//${url.host}`;
     }
 
     /**
      * Internal helper method to generate a request configuration (headers, parameters).
      *
      * @param transport optional transport request id
-     * @param testMode test mode enabled or not
+     * @param testMode optional url parameter to enable test mode
+     * @param safeMode optional url paramater to disable the safe model (safemode=false)
      * @returns the Axios response object for futher processing
      */
-    protected createConfig(transport?: string, testMode?: boolean): AxiosRequestConfig {
+    protected createConfig(transport?: string, testMode?: boolean, safeMode?: boolean): AxiosRequestConfig {
         const headers = {
             'Content-Type': 'application/atom+xml',
             type: 'entry',
             charset: 'UTF8'
         };
         const params: { [key: string]: string | boolean } = {
-            CodePage: "'UTF8'",
+            CodePage: `'UTF8'`,
             CondenseMessagesInHttpResponseHeader: 'X',
             format: 'json'
         };
@@ -169,21 +236,32 @@ export class Ui5AbapRepositoryService extends ODataService {
         if (testMode) {
             params.TestMode = true;
         }
+        if (safeMode !== undefined) {
+            params.SafeMode = safeMode;
+        }
 
-        return { headers, params };
+        // `axios` does not properly pass the default values of `maxBodyLength` and `maxContentLength`
+        // to `follow-redirects`: https://github.com/axios/axios/issues/4263
+        // Without this `follow-redirects` limits the max body length to 10MB and fails with the following
+        // message: “Request body larger than maxBodyLength limit”.
+        // Set both to infinity. It's the backend's responsibilty to reject messages sizes it cannot handle
+        const maxBodyLength = Infinity;
+        const maxContentLength = Infinity;
+
+        return { headers, params, maxBodyLength, maxContentLength };
     }
 
     /**
      * Create the request payload for a deploy request.
      *
-     * @param archive archive file path
+     * @param archive archive as buffer
      * @param name application name
      * @param description description for the deployed app
      * @param abapPackage ABAP package containing the app
      * @returns XML based request payload
      */
-    protected createPayload(archive: string, name: string, description: string, abapPackage: string): string {
-        const base64Data = readFileSync(archive, { encoding: 'base64' });
+    protected createPayload(archive: Buffer, name: string, description: string, abapPackage: string): string {
+        const base64Data = archive.toString('base64');
         const time = new Date().toISOString();
         const escapedName = encodeXmlValue(name);
         return (
@@ -238,9 +316,11 @@ export class Ui5AbapRepositoryService extends ODataService {
                 // We've nothing to return as we dont want to show the exception to the user!
                 return Promise.resolve(undefined);
             } else {
-                return isExisting
+                const response = isExisting
                     ? await this.put(`/Repositories('${encodeURIComponent(appName)}')`, payload, config)
                     : await this.post('/Repositories', payload, config);
+
+                return response;
             }
         } catch (error) {
             if (error?.response?.status === 504) {
@@ -293,10 +373,14 @@ export class Ui5AbapRepositoryService extends ODataService {
      * @param e.error error from Axios
      * @param e.host hostname
      */
-    protected logError({ error, host }: { error: AxiosError; host?: string }): void {
+    protected logError({ error, host }: { error: Error; host?: string }): void {
         this.log.error(error.message);
-        if (error.isAxiosError && error.response?.data?.['error']) {
-            prettyPrintError({ error: error.response.data['error'], host, log: this.log });
+        if (isAxiosError(error) && error.response?.data) {
+            if (error.response.data['error']) {
+                prettyPrintError({ error: error.response.data['error'], host, log: this.log });
+            } else {
+                this.log.error(error.response.data);
+            }
         }
     }
 }