+++ examples/simple-generator/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @sap-ux/generator-simple-fe
 
+## 0.2.59
+
+### Patch Changes
+
+-   Updated dependencies [a293cbb]
+    -   @sap-ux/fiori-elements-writer@0.8.5
+
+## 0.2.58
+
+### Patch Changes
+
+-   Updated dependencies [9b9b3d7]
+    -   @sap-ux/axios-extension@0.13.0
+
+## 0.2.57
+
+### Patch Changes
+
+-   Updated dependencies [116ff5e]
+    -   @sap-ux/axios-extension@0.12.0
+
+## 0.2.56
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@0.8.4
+-   @sap-ux/fiori-freestyle-writer@0.14.16
+
+## 0.2.55
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@0.8.3
+
+## 0.2.54
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@0.8.2
+-   @sap-ux/fiori-freestyle-writer@0.14.15
+
 ## 0.2.53
 
 ### Patch Changes