+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/environment-check
 
+## 0.7.2
+
+### Patch Changes
+
+-   Updated dependencies [9b9b3d7]
+    -   @sap-ux/axios-extension@0.13.0
+
+## 0.7.1
+
+### Patch Changes
+
+-   Updated dependencies [116ff5e]
+    -   @sap-ux/axios-extension@0.12.0
+
+## 0.7.0
+
+### Minor Changes
+
+-   63ad559: bump bas-sdk version
+
 ## 0.6.1
 
 ### Patch Changes